CLN: reorganize the structure of the library.
